refactor: extract abstract AgentStore base class
Move LRU caching, dedup loading, and eviction into an abstract
AgentStore base class. MainAgentStore and CodingAgentStore implement
the disk-dependent abstract methods (loadFromDisk, existsOnDisk,
deleteFromDisk, listSessionMetadata).
