+++ apps/backend/src/models/agent-store/agent-store.ts
@@ -1,0 +1,97 @@
+import type {SessionMetadata} from '@omnicraft/api-schema';
+
+import type {Agent} from '@/agent-core/agent/index.js';
+
+const MAX_CACHED_AGENTS = 50;
+
+interface CacheEntry {
+  agent: Agent;
+  lastAccessedAt: number;
+}
+
+/**
+ * Abstract base class for agent stores.
+ * Handles in-memory LRU caching and deduplication of concurrent loads.
+ * Subclasses implement disk persistence and session listing.
+ */
+export abstract class AgentStore {
+  private readonly cache = new Map<string, CacheEntry>();
+  private readonly loadingPromises = new Map<
+    string,
+    Promise<Agent | undefined>
+  >();
+
+  constructor(private readonly _sessionsDir: string) {}
+
+  get sessionsDir(): string {
+    return this._sessionsDir;
+  }
+
+  /** Registers an agent in the cache with LRU tracking. */
+  set(agent: Agent): void {
+    this.cache.set(agent.id, {agent, lastAccessedAt: Date.now()});
+    this.evictIfNeeded();
+  }
+
+  /** Retrieves an agent by id, loading from disk if not cached. */
+  async get(id: string): Promise<Agent | undefined> {
+    const entry = this.cache.get(id);
+    if (entry) {
+      entry.lastAccessedAt = Date.now();
+      return entry.agent;
+    }
+
+    const existing = this.loadingPromises.get(id);
+    if (existing) return existing;
+
+    const loadPromise = this.loadFromDisk(id);
+    this.loadingPromises.set(id, loadPromise);
+    try {
+      return await loadPromise;
+    } finally {
+      this.loadingPromises.delete(id);
+    }
+  }
+
+  /** Checks whether an agent exists in memory or on disk. Does not load. */
+  async has(id: string): Promise<boolean> {
+    if (this.cache.has(id)) return true;
+    return this.existsOnDisk(id);
+  }
+
+  /** Removes an agent from memory and deletes its session from disk. */
+  async delete(id: string): Promise<boolean> {
+    this.cache.delete(id);
+    return this.deleteFromDisk(id);
+  }
+
+  /** Lists persisted sessions with pagination. */
+  abstract listSessionMetadata(
+    offset: number,
+    limit: number,
+  ): Promise<{sessions: SessionMetadata[]; total: number}>;
+
+  /** Loads an agent from disk. Returns undefined if not found. */
+  protected abstract loadFromDisk(id: string): Promise<Agent | undefined>;
+
+  /** Checks whether an agent session exists on disk. */
+  protected abstract existsOnDisk(id: string): Promise<boolean>;
+
+  /** Deletes an agent session from disk. */
+  protected abstract deleteFromDisk(id: string): Promise<boolean>;
+
+  private evictIfNeeded(): void {
+    if (this.cache.size <= MAX_CACHED_AGENTS) return;
+
+    const entries = [...this.cache.entries()]
+      .filter(
+        ([, e]) => !e.agent.isRunning && e.agent.sseLog.activeReaderCount === 0,
+      )
+      .sort((a, b) => a[1].lastAccessedAt - b[1].lastAccessedAt);
+
+    for (const [id] of entries) {
+      if (this.cache.size <= MAX_CACHED_AGENTS) break;
+      this.cache.delete(id);
+    }
+  }
+}
