+++ apps/backend/src/models/agent-store/index.ts
@@ -1,2 +1,3 @@
+export {AgentStore} from './agent-store.js';
 export {CodingAgentStore} from './coding-agent-store.js';
 export {MainAgentStore} from './main-agent-store.js';
