+++ apps/backend/src/models/agent-store/main-agent-store.ts
@@ -14,21 +14,10 @@ import {agentEventBus} from '@/agent-core/events/index.js';
 import {isFileNotFoundError} from '@/helpers/fs.js';
 import {logger} from '@/logger.js';
 
-const MAX_CACHED_AGENTS = 50;
+import {AgentStore} from './agent-store.js';
 
-interface CacheEntry {
-  agent: Agent;
-  lastAccessedAt: number;
-}
-
-export class MainAgentStore {
+export class MainAgentStore extends AgentStore {
   private static instance: MainAgentStore | null = null;
-  private readonly cache = new Map<string, CacheEntry>();
-  private readonly loadingPromises = new Map<
-    string,
-    Promise<Agent | undefined>
-  >();
-  private readonly _sessionsDir: string;
 
   private readonly onAgentCreated = (agent: Agent): void => {
     if (agent instanceof MainAgent) {
@@ -37,11 +26,7 @@ export class MainAgentStore {
   };
 
   private constructor(sessionsDir: string) {
-    this._sessionsDir = sessionsDir;
-  }
-
-  get sessionsDir(): string {
-    return this._sessionsDir;
+    super(sessionsDir);
   }
 
   /** Returns the singleton instance. */
@@ -76,63 +61,13 @@ export class MainAgentStore {
     MainAgentStore.instance = null;
   }
 
-  /** Registers an agent in the cache with LRU tracking. */
-  set(agent: Agent): void {
-    this.cache.set(agent.id, {agent, lastAccessedAt: Date.now()});
-    this.evictIfNeeded();
-  }
-
-  /** Retrieves an agent by id, loading from disk if not cached. */
-  async get(id: string): Promise<Agent | undefined> {
-    const entry = this.cache.get(id);
-    if (entry) {
-      entry.lastAccessedAt = Date.now();
-      return entry.agent;
-    }
-
-    const existing = this.loadingPromises.get(id);
-    if (existing) return existing;
-
-    const loadPromise = this.loadFromDisk(id);
-    this.loadingPromises.set(id, loadPromise);
-    try {
-      return await loadPromise;
-    } finally {
-      this.loadingPromises.delete(id);
-    }
-  }
-
-  /** Checks whether an agent exists in memory or on disk. Does not load. */
-  async has(id: string): Promise<boolean> {
-    if (this.cache.has(id)) return true;
-    return this.existsOnDisk(id);
-  }
-
-  /** Removes an agent from memory and deletes its session directory. */
-  async delete(id: string): Promise<boolean> {
-    this.cache.delete(id);
-    const sessionDir = path.join(this._sessionsDir, id);
-    try {
-      await rm(sessionDir, {recursive: true, force: true});
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Scans the sessions directory and returns paginated session metadata.
-   *
-   * First stats all snapshot files to obtain mtime for sorting, then
-   * only reads snapshot content for entries in the requested page range.
-   */
   async listSessionMetadata(
     offset: number,
     limit: number,
   ): Promise<{sessions: SessionMetadata[]; total: number}> {
     let entries: string[];
     try {
-      entries = await readdir(this._sessionsDir);
+      entries = await readdir(this.sessionsDir);
     } catch {
       return {sessions: [], total: 0};
     }
@@ -143,7 +78,7 @@ export class MainAgentStore {
       entries.map(async (entry) => {
         try {
           const fileStat = await stat(
-            agentPersistence.snapshotPath(this._sessionsDir, entry),
+            agentPersistence.snapshotPath(this.sessionsDir, entry),
           );
           statResults.push({id: entry, mtime: fileStat.mtimeMs});
         } catch (e) {
@@ -178,6 +113,30 @@ export class MainAgentStore {
     return {sessions, total};
   }
 
+  protected async loadFromDisk(id: string): Promise<Agent | undefined> {
+    if (!(await this.existsOnDisk(id))) return undefined;
+    return MainAgent.restore(this.sessionsDir, id);
+  }
+
+  protected async existsOnDisk(id: string): Promise<boolean> {
+    try {
+      await access(agentPersistence.snapshotPath(this.sessionsDir, id));
+      return true;
+    } catch {
+      return false;
+    }
+  }
+
+  protected async deleteFromDisk(id: string): Promise<boolean> {
+    const sessionDir = path.join(this.sessionsDir, id);
+    try {
+      await rm(sessionDir, {recursive: true, force: true});
+      return true;
+    } catch {
+      return false;
+    }
+  }
+
   /**
    * Reads session metadata from metadata.json, falling back to snapshot.json
    * for sessions created before the sidecar file was introduced.
@@ -185,49 +144,17 @@ export class MainAgentStore {
   private async readSessionMetadataFile(id: string): Promise<string> {
     try {
       return await readFile(
-        agentPersistence.metadataPath(this._sessionsDir, id),
+        agentPersistence.metadataPath(this.sessionsDir, id),
         'utf-8',
       );
     } catch (error: unknown) {
       if (isFileNotFoundError(error)) {
         return readFile(
-          agentPersistence.snapshotPath(this._sessionsDir, id),
+          agentPersistence.snapshotPath(this.sessionsDir, id),
           'utf-8',
         );
       }
       throw error;
-    }
-  }
-
-  private async loadFromDisk(id: string): Promise<Agent | undefined> {
-    if (!(await this.existsOnDisk(id))) return undefined;
-    const agent = await MainAgent.restore(this._sessionsDir, id);
-    const entry = this.cache.get(id);
-    if (entry) entry.lastAccessedAt = Date.now();
-    return agent;
-  }
-
-  private async existsOnDisk(id: string): Promise<boolean> {
-    try {
-      await access(agentPersistence.snapshotPath(this._sessionsDir, id));
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  private evictIfNeeded(): void {
-    if (this.cache.size <= MAX_CACHED_AGENTS) return;
-
-    const entries = [...this.cache.entries()]
-      .filter(
-        ([, e]) => !e.agent.isRunning && e.agent.sseLog.activeReaderCount === 0,
-      )
-      .sort((a, b) => a[1].lastAccessedAt - b[1].lastAccessedAt);
-
-    for (const [id] of entries) {
-      if (this.cache.size <= MAX_CACHED_AGENTS) break;
-      this.cache.delete(id);
     }
   }
 }
