+++ apps/backend/src/services/agent-session/agent-session-service.ts
@@ -12,7 +12,11 @@ import type {SseEvent} from '@omnicraft/sse-events';
 import {CodingAgent, MainAgent} from '@/agent/agents/index.js';
 import type {AgentSseLogReaderOptions} from '@/agent-core/agent/agent-sse-log.js';
 import type {Agent} from '@/agent-core/agent/index.js';
-import {CodingAgentStore, MainAgentStore} from '@/models/agent-store/index.js';
+import {
+  AgentStore,
+  CodingAgentStore,
+  MainAgentStore,
+} from '@/models/agent-store/index.js';
 import {SettingsManager} from '@/models/settings-manager/index.js';
 
 import {getLlmConfig} from './helpers.js';
@@ -20,7 +24,7 @@ import type {CreateSessionResult} from './types.js';
 import {CreateSessionError} from './types.js';
 import {validateSessionPaths} from './validation.js';
 
-function getStore(agentType: AgentType): MainAgentStore | CodingAgentStore {
+function getStore(agentType: AgentType): AgentStore {
   switch (agentType) {
     case AgentType.CHAT:
       return MainAgentStore.getInstance();
